pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ['py38']
line-length = 100
skip-string-normalization = true

# [tool.nitpick]
# style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "pyentrez"
description = "Terminal-based TUI suite of Biopython's Entrez tools."
version = "0.1.0"
license = "MIT"
authors = ["William Slattery <slatte26@msu.edu>"]
readme = "README.rst"
repository = "https://github.com/BGASM/pyEntrez"
documentation = "https://pyentrez.readthedocs.io/en/latest/"

keywords = [
  "Biopython",
  "Pubmed",
  "NCBI",
  "Abstract",
  "Entrez"
]

classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Environment :: Console :: Curses",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Healthcare Industry",
  "Intended Audience :: End Users/Desktop",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering",
  "Natural Language :: English"
]

[tool.poetry.dependencies]
python = ">=3.6,<3.9"
ConfigArgParse = "^1.2.3"
loguru = "^0.5.3"
py-cui = "^0.1.2"
bio = "^0.1.9"
PyYAML = "^5.3.1"

# Extras
pymongo = { version = "^3.11.2", optional = true }
black = { version = "^20.8b1", optional = true }
hunter = { version = "^3.3.1", optional = true }

# The following are for NLP use and are optional.
scispacy = { version = "^0.3.0", optional = true }
tldextract = { version = "^3.1.0", optional = true }
nmslib = { version = "^2.0.6", optional = true }
en-core-sci-lg = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.3.0/en_core_sci_lg-0.3.0.tar.gz", optional = true}
lorem = {version = "^0.1.1", optional = true}

[tool.poetry.extras]
nlp = ["scispacy", "tldextract", "nmslib", "en_core_sci_lg"]
lint = ["mypy", "wemake-python-styleguide", "flake8-pytest-style", "black", "hunter"]
mongo = ["pymongo"]
lorem = ["lorem"]

[tool.poetry.dev-dependencies]
nitpick = "^0.23"
safety = "^1.9"
pytest = "^5.4"
pytest-cov = "^2.10"
pytest-randomly = "^3.5"
sphinx = "^3.3"
sphinx-autodoc-typehints = "^1.11"
doc8 = "^0.8"
m2r2 = "^0.2"
tomlkit = "^0.7"
lxml = "^4.6.2"
MonkeyType = "^20.5.0"
yappi = "^1.3.2"
sphinx-rtd-theme = "^0.5.0"
pytest-mock = "^3.3.1"
# Linters
mypy = "^0.790"
wemake-python-styleguide = "^0.14"
flake8-pytest-style = "^1.3"
attrs = "^20.3.0"








